requirements.txt
bs4
pytesseract
fastapi
requests
pandas
lxml
uvicorn
httpx
psycopg2-binary
python-dotenv
schedule
